refactor: extract create_index module from table.rs (#2949)
Split index creation and management code from table.rs into a new
private table/create_index.rs submodule, completing the last proposed
module split from issue #2949.

Moved NativeTable methods: load_indices, validate_index_type,
build_ivf_params, get_num_sub_vectors, get_vector_dimension,
resolve_index_field, make_index_params, get_index_type_for_field.
BaseTable trait impl index methods remain in table.rs as delegates.

11 index-related tests moved to the new module. 4 tests converted
from tempdir() to memory:// URIs, with record_batch!() macro where
applicable. Simplified a dead branch in get_index_type_for_field.

table.rs reduced from 5009 to 3804 lines.
